recommendation_service/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy>=2.0




Chạy lệnh pip install -r requirements.txt 